src/requirements.txt
fastapi
pydantic
uvicorn
langchain
langchain-openai
requests
praw
python-dotenv
chromadb
pydantic-settings
openai
